fix security alarms

+++ services/bundlemgr/src/base_bundle_installer.cpp
@@ -5926,7 +5926,7 @@ ErrCode BaseBundleInstaller::ParseHapPaths(const InstallParam &installParam,
         FileStat fileStat;
         ErrCode res = InstalldClient::GetInstance()->GetFileStat(bundleNameDir, fileStat);
         int32_t sharedMode = S_IRWXU | S_IRWXG | S_ISGID;
-        if (res == ERR_OK && (fileStat.mode & ServiceConstants::MODE_BASE) != sharedMode) {
+        if (res == ERR_OK && (static_cast<uint32_t>(fileStat.mode) & ServiceConstants::MODE_BASE) != sharedMode) {
             LOG_W(BMS_TAG_INSTALLER, "shared dir mode is not correct %{public}d for %{public}s",
                 fileStat.mode, bundleNameDir.c_str());
             fileStat.mode = sharedMode;
